requirements.txt
openai>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
praw>=7.7.1
tenacity>=8.2.3
numpy>=1.26.0
scikit-learn>=1.4.0
slack_bolt>=1.20.0
slack_sdk>=3.21.3
Flask>=3.0.0
